Moved FunSuite related tests to jvm/funsuite-test.
