+++ jvm/funsuite-test/src/test/scala/org/scalatest/funsuite/AsyncFunSuiteSpec.scala
@@ -13,24 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.funsuite
 
-import org.scalatest.SharedHelpers.EventRecordingReporter
+import org.scalatest._
+import org.scalatest.SharedHelpers.{EventRecordingReporter, thisLineNumber}
 import scala.concurrent.{Promise, ExecutionContext, Future}
 import org.scalatest.concurrent.SleepHelper
 import org.scalatest.events.{InfoProvided, MarkupProvided}
+import org.scalatest.exceptions.DuplicateTestNameException
 
 import scala.util.Success
 import org.scalatest.funspec.AnyFunSpec
-import org.scalatest.funsuite.AsyncFunSuiteLike
+import org.scalatest.funsuite.AsyncFunSuite
 
-class AsyncFunSuiteLikeSpec extends AnyFunSpec {
+class AsyncFunSuiteSpec extends AnyFunSpec {
 
-  describe("AsyncFunSuiteLike") {
+  describe("AsyncFunSuite") {
 
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSuite extends AsyncFunSuiteLike with ParallelTestExecution {
+      class ExampleSuite extends AsyncFunSuite with ParallelTestExecution {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -69,6 +71,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
         override def newInstance = new ExampleSuite
       }
 
+
       val rep = new EventRecordingReporter
       val suite = new ExampleSuite
       val status = suite.run(None, Args(reporter = rep))
@@ -90,7 +93,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSuite extends AsyncFunSuiteLike with ParallelTestExecution {
+      class ExampleSuite extends AsyncFunSuite with ParallelTestExecution {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -142,7 +145,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSuite extends AsyncFunSuiteLike {
+      class ExampleSuite extends AsyncFunSuite {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -188,7 +191,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSuite extends AsyncFunSuiteLike {
+      class ExampleSuite extends AsyncFunSuite {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -232,7 +235,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends AsyncFunSuiteLike {
+      class ExampleSpec extends AsyncFunSuite {
 
         test("test 1") {
           Future {
@@ -272,7 +275,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends AsyncFunSuiteLike {
+      class ExampleSpec extends AsyncFunSuite {
 
         test("test 1") {
           val promise = Promise[Assertion]
@@ -328,7 +331,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends AsyncFunSuiteLike {
+      class ExampleSpec extends AsyncFunSuite {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -356,7 +359,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends AsyncFunSuiteLike {
+      class ExampleSpec extends AsyncFunSuite {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -401,7 +404,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends AsyncFunSuiteLike {
+      class ExampleSpec extends AsyncFunSuite {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -439,7 +442,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
         info(
           "hi there"
         )
@@ -458,7 +461,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -487,7 +490,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in Future returned by test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -518,7 +521,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
         note(
           "hi there"
         )
@@ -537,7 +540,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -559,7 +562,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -583,7 +586,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
         alert(
           "hi there"
         )
@@ -602,7 +605,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -624,7 +627,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -648,7 +651,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
         markup(
           "hi there"
         )
@@ -667,7 +670,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -696,7 +699,7 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by test body") {
-      class MySuite extends AsyncFunSuiteLike  {
+      class MySuite extends AsyncFunSuite  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 
@@ -726,8 +729,34 @@ class AsyncFunSuiteLikeSpec extends AnyFunSpec {
       assert(markupProvided.text == "hi there")
     }
 
+    it("should generate a DuplicateTestNameException when duplicate test name is detected") {
+      class TestSpec extends AsyncFunSuite {
+        test("test 1") { succeed }
+        test("test 1") { succeed }
+      }
+      val e = intercept[DuplicateTestNameException] {
+        new TestSpec
+      }
+      assert("AsyncFunSuiteSpec.scala" == e.failedCodeFileName.get)
+      assert(e.failedCodeLineNumber.get == thisLineNumber - 6)
+      assert(!e.cause.isDefined)
+    }
+
+    it("should generate a DuplicateTestNameException when duplicate test name is detected using ignore") {
+      class TestSpec extends AsyncFunSuite {
+        test("test 1") { succeed }
+        ignore("test 1") { succeed }
+      }
+      val e = intercept[DuplicateTestNameException] {
+        new TestSpec
+      }
+      assert("AsyncFunSuiteSpec.scala" == e.failedCodeFileName.get)
+      assert(e.failedCodeLineNumber.get == thisLineNumber - 6)
+      assert(!e.cause.isDefined)
+    }
+
     it("should allow other execution context to be used") {
-      class TestSpec extends AsyncFunSuiteLike {
+      class TestSpec extends AsyncFunSuite {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
