+++ jvm/funsuite-test/src/test/scala/org/scalatest/funsuite/FunSuiteSuite.scala
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.funsuite
 
-import SharedHelpers._
+import org.scalatest._
+import org.scalatest.SharedHelpers._
 import org.scalatest.events._
 import org.scalatest.exceptions.DuplicateTestNameException
 import org.scalatest.exceptions.TestRegistrationClosedException
